"Fixed" parenting issue for anonymous elements
I say fixed in quotes, because it is really hacky and there are
glaring bugs when you hit the issue. I need to debug more
into the platform to figure out the issue, i'm guessing it is
something in my code. However, this replaces a crashing bug
with visual artifacts, so it is a move in the right direction.